requirements.txt
streamlit
colorthief
requests
Pillow